[Feat] #49 - GIFImageView 컴포넌트 분리
