+++ SoBunSoBun/Common/Components/BlueMeatballs/BlueMeatballsRefreshController.swift
@@ -12,19 +12,14 @@ import RxSwift
 import OSLog
 
 class BlueMeatballsRefreshController: UIRefreshControl {
-    private let logger = Logger(
-        subsystem: "SoBunSoBun",
-        category: "BlueMeatballsRefreshController"
-    )
-    
     private let disposeBag = DisposeBag()
     
-    private let gifImageView: UIImageView = {
-        let iv = UIImageView()
-        iv.contentMode = .scaleAspectFit
-        iv.clipsToBounds = true
+    private let gifImageView: GIFImageView = {
+        let gif = GIFImageView(fileName: "BlueMeatballs", speed: 5)
+        gif.contentMode = .scaleAspectFit
+        gif.clipsToBounds = true
         
-        return iv
+        return gif
     }()
     
     override init(frame: CGRect = .zero) {
@@ -38,39 +33,7 @@ class BlueMeatballsRefreshController: UIRefreshControl {
     }
     
     private func configureUI() {
-        tintColor = .clear
-        
-        // gif 적용
-        guard let gifURL = Bundle.main.url(forResource: "BlueMeatballs", withExtension: "gif"),
-              let gifData = try? Data(contentsOf: gifURL),
-              let source = CGImageSourceCreateWithData(gifData as CFData, nil) else {
-            logger.critical("BlueMeatballs GIF 파일을 찾지 못함")
-            return
-        }
-        
-        let frameCount = CGImageSourceGetCount(source)
-        var images = [UIImage]()
-        var totalDuration: TimeInterval = 0
-        
-        for i in 0..<frameCount {
-            if let cgImage = CGImageSourceCreateImageAtIndex(source, i, nil) {
-                let image = UIImage(cgImage: cgImage)
-                images.append(image)
-                
-                // frame duration 가져오기
-                if let properties = CGImageSourceCopyPropertiesAtIndex(source, i, nil) as? [String: Any],
-                   let gifInfo = properties[kCGImagePropertyGIFDictionary as String] as? [String: Any],
-                   let duration = gifInfo[kCGImagePropertyGIFDelayTime as String] as? TimeInterval {
-                    totalDuration += duration
-                } else {
-                    totalDuration += 0.1 // 기본값
-                }
-            }
-        }
-        
-        gifImageView.animationImages = images
-        gifImageView.animationDuration = totalDuration * (1 / 5) // 기존 gif 속도의 5배
-        gifImageView.animationRepeatCount = 0
+        self.tintColor = .clear
         
         addSubview(gifImageView)
